chore: move cases related methods to 'case.ts' file
